[Chore] #73 - homeVC 수정
